+++ app/src/main/java/com/sopt/now/MyApplication.kt
@@ -1,7 +1,6 @@
 package com.sopt.now
 
 import android.app.Application
-import com.sopt.now.data.PreferenceUtil
 import dagger.hilt.android.HiltAndroidApp
 
 @HiltAndroidApp
