+++ app/src/main/java/com/sopt/now/data/model/remote/ResponseSignUpDto.kt
@@ -1,15 +1,12 @@
-package com.sopt.now.data
+package com.sopt.now.data.model.remote
 
-import com.sopt.now.presentation.model.UserInfo
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class ResponseUserInfoDto(
+data class ResponseSignUpDto(
     @SerialName("code")
     val code: Int,
     @SerialName("message")
     val message: String,
-    @SerialName("data")
-    val data: UserInfo,
 )
