+++ app/src/main/java/com/sopt/now/data/model/remote/ResponseUserInfoDto.kt
@@ -1,12 +1,15 @@
-package com.sopt.now.data
+package com.sopt.now.data.model.remote
 
+import com.sopt.now.presentation.model.UserInfo
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class ResponseSignInDto(
+data class ResponseUserInfoDto(
     @SerialName("code")
     val code: Int,
     @SerialName("message")
     val message: String,
+    @SerialName("data")
+    val data: UserInfo,
 )
