+++ app/src/main/java/com/sopt/now/data/model/service/AuthService.kt
@@ -1,5 +1,10 @@
-package com.sopt.now.data
+package com.sopt.now.data.model.service
 
+import com.sopt.now.data.model.remote.RequestSignInDto
+import com.sopt.now.data.model.remote.RequestSignUpDto
+import com.sopt.now.data.model.remote.ResponseSignInDto
+import com.sopt.now.data.model.remote.ResponseSignUpDto
+import com.sopt.now.data.model.remote.ResponseUserInfoDto
 import retrofit2.Call
 import retrofit2.http.Body
 import retrofit2.http.GET
