+++ app/src/main/java/com/sopt/now/presentation/main/friends/FriendViewModel.kt
@@ -5,8 +5,8 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.sopt.now.data.database.FriendDatabase
-import com.sopt.now.data.model.Friend
-import com.sopt.now.data.model.FriendEntity
+import com.sopt.now.data.model.local.Friend
+import com.sopt.now.data.model.local.FriendEntity
 import kotlinx.coroutines.launch
 
 class FriendViewModel(private val friendDatabase: FriendDatabase) : ViewModel() {
