+++ app/src/main/java/com/sopt/now/presentation/main/mypage/MyPageViewModel.kt
@@ -4,8 +4,8 @@ import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import com.sopt.now.data.PreferenceUtil
-import com.sopt.now.data.ResponseUserInfoDto
+import com.sopt.now.data.SharedPreferenceUtil
+import com.sopt.now.data.model.remote.ResponseUserInfoDto
 import com.sopt.now.data.ServicePool
 import com.sopt.now.presentation.model.UserInfo
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -15,7 +15,7 @@ import retrofit2.Response
 import javax.inject.Inject
 
 @HiltViewModel
-class MyPageViewModel @Inject constructor(private val sharedPreferences: PreferenceUtil) :
+class MyPageViewModel @Inject constructor(private val sharedPreferences: SharedPreferenceUtil) :
     ViewModel() {
     private var _userInfo: MutableLiveData<UserInfo> = MutableLiveData(UserInfo("", "", ""))
     val userInfo: LiveData<UserInfo> get() = _userInfo
